Fix issue 234: machine readable format for asetekpro pump
This commit adds support for the --machine option to the asetekpro pump
so that it produces machine readable output for the values of the
temperature, fans and pump.

The fix was to simply add calls to the msg_machine() method similar to
what already exists in the hydro_asetek.c file.

I have tested this with a H150i Pro and it produces output like:

temperature:0:33.70
fan:0:0:604:0
fan:1:0:618:0
fan:2:0:600:0
pump:2:2820:0

This change should work with other devices that use the asetekpro pump